% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filters Data}
\usage{
filter_data(data)
}
\arguments{
\item{data}{Dataset to filter.}
}
\value{
Filtered and cleaned dataset.
}
\description{
Filters Data
}
\examples{
filter_data()
}
