% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Loading in Measles Data from TidyTuesday}
\usage{
load_data()
}
\value{
A list containing two dataframes that record measles incidence at different
levels of time. The first, cases_month, contains monthly data at the country-year-month
level (i.e. Algeria, January 2012). The second, cases_year, contains yearly data
at the country-year level (i.e. Algeria, 2012). Together, these datasets include
information for 194 countries across multiple years and months.
}
\description{
Loading in Measles Data from TidyTuesday
}
\examples{
load_data()
}
