% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_WorldData.R
\name{load_WorldData}
\alias{load_WorldData}
\title{Get World Data for Map}
\usage{
load_WorldData(target_year)
}
\value{
Cleaned Dataset with Map Polygons for Leaflet Map
}
\description{
Get World Data for Map
}
\examples{
load_WorldData(2016)
}
