% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCases.R
\name{mapCases}
\alias{mapCases}
\title{Measles Cases Leaflet Map}
\usage{
mapCases(target_year)
}
\arguments{
\item{year}{Year}
}
\value{
Generates interactive leaflet that covers countries' measles cases,
MCV1 Coverage and Nominal GDP per Capita for a given year.
}
\description{
Measles Cases Leaflet Map
}
\examples{
mapCases(2015)
}
