% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCases.R
\name{modelCases}
\alias{modelCases}
\title{Produces NB GLM statistical analysis output for data}
\usage{
modelCases(interaction = T)
}
\arguments{
\item{interaction}{whether to include interaction between Region and other variables}
}
\value{
Data frame with NB GLM coefficients
}
\description{
Produces NB GLM statistical analysis output for data
}
\examples{
modelCases()
}
