% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVaccination.R
\name{plotVaccination}
\alias{plotVaccination}
\title{Create Line Plot Comparing MCV1 Coverage}
\usage{
plotVaccination(country, min_year = 2014, max_year = 2024)
}
\arguments{
\item{country}{The country to be analyzed.}

\item{min_year}{The minimum year for analysis. Note: Earliest data on
file is 2012 for most regions except the Americas that start at 2014.}

\item{max_year}{The maximum year for analysis. Note: Latest data on
file is 2025.}
}
\value{
Generates a line plot displaying a country's MCV1 vaccination rate across time.
}
\description{
Create Line Plot Comparing MCV1 Coverage
}
\examples{
plotVaccination(country = "Romania")
}
