% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeYear.R
\name{summarizeYear}
\alias{summarizeYear}
\title{Summarizes a years' worth of global Measles case data}
\usage{
summarizeYear(year)
}
\arguments{
\item{year}{Year to summarize}
}
\value{
Data frame with summarized Measles case data for a range of years
}
\description{
Summarizes a years' worth of global Measles case data
}
\examples{
summarizeYear(2015)
}
