% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRegion.R
\name{validateRegion}
\alias{validateRegion}
\title{Validates a region}
\usage{
validateRegion(region, data)
}
\arguments{
\item{region}{Region to be checked for.}

\item{data}{Data frame which may include the region.}
}
\value{
Invisible TRUE if test is passed.
}
\description{
Validates a region
}
