% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateYear.R
\name{validateYear}
\alias{validateYear}
\title{Validates if a Year or Pair of Years Can Be Used (Found in Data)}
\usage{
validateYear(year1, year2 = NULL)
}
\arguments{
\item{year1}{Year to be tested. If in a pair, this in the minimum year to be tested.}

\item{year2}{Max year to be tested.}
}
\value{
Invisible TRUE if all tests are passed.
}
\description{
Validates if a Year or Pair of Years Can Be Used (Found in Data)
}
\examples{
validateYear(2015)
}
