% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCases.R
\name{plotCases}
\alias{plotCases}
\title{Plot number of cases in region(s) by year.}
\usage{
plotCases(region, legend = F)
}
\arguments{
\item{region}{A vector of region names.}

\item{legend}{Whether a legend should be displayed (warning: typically illegible)}
}
\value{
A plot displaying number of measles cases across all years in these regions.
}
\description{
Plot number of cases in region(s) by year.
}
\examples{
plotCases("AMRO")
}
