% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRegion.R
\name{summarizeRegion}
\alias{summarizeRegion}
\title{Summarizes the cases within a region's countries}
\usage{
summarizeRegion(region)
}
\arguments{
\item{region}{Region to summarize. Valid regions include "Africa," "Americas," "Eastern Mediterranean," "Europe," "South East Asia," and "Western Pacific"}
}
\value{
Data frame with summarized Measles case data for a single region
}
\description{
Summarizes the cases within a region's countries
}
\examples{
summarizeRegion("Americas")
}
