% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRegion.R
\name{summarizeRegion}
\alias{summarizeRegion}
\title{Summarizes a region's countries within the data set}
\usage{
summarizeRegion(region)
}
\arguments{
\item{region}{Region to summarize}
}
\value{
Data frame with summarized Measles case data for a single region
}
\description{
Summarizes a region's countries within the data set
}
\examples{
summarizeRegion(AMRO)
}
